refactor: rename useStorageListener to use-storage-listener for consistency
